♻️ MyWork 등록 수정 되라 얍

+++ src/main/java/com/MoneyPlant/dto/MyWorkDto.java
@@ -12,7 +12,7 @@ public class MyWorkDto {
     private String myPayType;
     private int myWorkMoney;
     private int myWorkTime;
-    private int myWorkTax;
+    private double myWorkTax;
     private int myWorkPay;
     private String myPayDay;
